Add standlone file
This file has everything in it, so you don't need to mess around with sourcing
the colors file separately.  You can start by just sourcing this file and then
editing it.  (The thing you will most likely need to edit is the last line,
which sets your prompt string directly.)